+++ Plugins/Root/include/ACTFW/Plugins/Root/RootPlanarClusterWriter.hpp
@@ -98,7 +98,7 @@ namespace Root {
     std::vector<float> m_t_gz;       ///< truth position global z
     std::vector<float> m_t_lx;       ///< truth position local x
     std::vector<float> m_t_ly;       ///< truth position local y
-    std::vector<int>   m_t_barcode;  ///< associated truth particle barcode
+    std::vector<unsigned long>   m_t_barcode;  ///< associated truth particle barcode
   };
 
 }  // namespace Root
